Prevent onItemSelectedHandler on save/restore
The onItemSelectedListener gets called asynchronously after
saved state has been applied, resulting in the unit spinner to be
reset to default.

There seems to be no way to prevent the handler from running for non-user
initiated state change of the spinner.

Work around this by only changing the unit spinner if the current pos
is different from what has been set from saved state. This effectively
disabled the handler as long as the selected category is not changed.